Fix tiles build.
git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@4881 c06c8d41-db1a-0410-9941-cceddc491573

+++ crawl-ref/source/tile1.cc
@@ -300,7 +300,7 @@ static int _tileidx_monster_base(int mon_idx, bool detected)
 
     case MONS_HYDRA:
         // Number of heads
-        return TILE_MONS_HYDRA + std::min(mon->number, 7) - 1;
+        return TILE_MONS_HYDRA + std::min((int)mon->number, 7) - 1;
 
     case MONS_SKELETON_SMALL:
         return TILE_MONS_SKELETON_SMALL;
